requirements.txt
openai-whisper
sounddevice
python-xlib
numpy
